--- backend/package.json
@@ -1,30 +1,0 @@
-{
-    "scripts": {
-        "build": "tsc",
-        "start": "node dist/server.js",
-        "dev": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/server.ts",
-        "lint": "eslint 'src/**/*.ts'"
-    },
-    "devDependencies": {
-        "@types/cors": "^2.8.18",
-        "@types/express": "^5.0.2",
-        "@types/lodash": "^4.17.16",
-        "@types/luxon": "^3.6.2",
-        "@types/node": "^22.15.18",
-        "dotenv": "^16.5.0",
-        "eslint": "^9.27.0",
-        "express": "^5.1.0",
-        "install": "^0.13.0",
-        "nodemon": "^3.1.10",
-        "npm": "^11.4.0",
-        "prettier": "^3.5.3",
-        "ts-node": "^10.9.2",
-        "tsconfig-paths": "^4.2.0",
-        "typescript": "^5.8.3"
-    },
-    "dependencies": {
-        "cors": "^2.8.5",
-        "lodash": "^4.17.21",
-        "luxon": "^3.6.1"
-    }
-}
