--- backend/src/config/config.ts
@@ -1,15 +1,0 @@
-import dotenv from "dotenv";
-
-dotenv.config();
-
-interface Config {
-    port: number;
-    nodeEnv: string;
-}
-
-const config: Config = {
-    port: Number(process.env.PORT) || 3000,
-    nodeEnv: process.env.NODE_ENV || "development",
-};
-
-export default config;
