--- backend/src/controllers/playlistController.ts
@@ -1,29 +1,0 @@
-import { isNull } from "lodash";
-import { Playlist } from "@models/playlist";
-import { Request, Response, NextFunction } from "express";
-
-export const getPlaylistById = async (
-    req: Request,
-    res: Response,
-    next: NextFunction
-) => {
-    try {
-        const playlist = new Playlist({ id: req.params.playlistId });
-
-        await playlist.create();
-
-        if (isNull(playlist.details)) {
-            res.status(400).json({ message: "No playlist found on this id" });
-
-            return;
-        }
-
-        res.status(200).json({
-            details: playlist.details,
-            imgUrl: playlist.details?.snippet.thumbnails["maxres"].url,
-            totalLength: await playlist.getPlaylistDuration(),
-        });
-    } catch (error) {
-        next(error);
-    }
-};
