--- backend/src/server.ts
@@ -1,6 +1,0 @@
-import app from "./app";
-import config from "./config/config";
-
-app.listen(config.port, () => {
-    console.log(`Server running on port ${config.port}`);
-});
