--- frontend/src/types/index.d.ts
@@ -1,1 +1,0 @@
-type Nullable<T> = T | null | undefined;
